requirements.txt
matplotlib==3.4.2
nibabel=3.2.1
pandas==1.2.5
pytest=6.2.4
requests==2.25.1
scikit-learn==0.24.2
tensorflow-gpu==2.5.0
tqdm==4.61.1
xlrd==2.0.1